+++ src/Language/Haskell/Names.hs
@@ -1,14 +1,18 @@
 module Language.Haskell.Names
   (
-  -- * Core functions
-    computeInterfaces
-  , getInterfaces
-  , annotateModule
+  -- * Functions
+    resolve
+  , annotate
   -- * Types
+  , Environment
   , Symbol(..)
   , Scoped(..)
   , NameInfo(..)
   , Error(..)
+  -- * Reading and writing environments
+  , readSymbols
+  , writeSymbols
+  , loadBase
   -- * Pretty printing
   , ppError
   , ppSymbol
@@ -16,3 +20,5 @@ module Language.Haskell.Names
 
 import Language.Haskell.Names.Types
 import Language.Haskell.Names.Recursive
+import Language.Haskell.Names.Environment
+